% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{api_client}
\alias{api_client}
\title{API client}
\usage{
api_client(base_url = Sys.getenv("GH_API_URL"), port = NULL)
}
\arguments{
\item{base_url}{character scalar; API base url}

\item{port}{numeric scalar; port}
}
\description{
API client
}
