% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{get_route}
\alias{get_route}
\title{Get the route for a given start and end point.}
\usage{
get_route(from_lnglat, to_lnglat, ...)
}
\arguments{
\item{from_lnglat}{numeric vector; coordinates [lng, lat] of start point}

\item{to_lnglat}{numeric vector; coordinates [lng, lat] of end point}

\item{...}{additional query parameters}
}
\description{
Get the route for a given start and end point.
}
