% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt-endpoint.R
\name{gh_available_spt_columns}
\alias{gh_available_spt_columns}
\title{Get a vector with available columns of the spt endpoint}
\usage{
gh_available_spt_columns()
}
\description{
Get a vector with available columns of the spt endpoint
}
