% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-route.R
\name{gh_bbox}
\alias{gh_bbox}
\alias{gh_bbox.gh_route}
\alias{gh_bbox.gh_info}
\title{Extract the bounding box from a gh object}
\usage{
gh_bbox(data)

\method{gh_bbox}{gh_route}(data)

\method{gh_bbox}{gh_info}(data)
}
\arguments{
\item{data}{A \code{gh_route} or \code{gh_info} object.}
}
\description{
Extract the bounding box from a gh object
}
