% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-endpoint.R
\name{gh_get_info}
\alias{gh_get_info}
\title{Get information about the GraphHooper instance}
\usage{
gh_get_info()
}
\description{
Get information about the GraphHooper instance
}
\examples{
\dontrun{
  gh_get_info()$data_date

  gh_get_info()$version

  gh_get_info() \%>\%
    gh_bbox()
}
}
