% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt-endpoint.R
\name{gh_get_spt}
\alias{gh_get_spt}
\title{Get the shortest path tree for a given start point}
\usage{
gh_get_spt(start_point, time_limit = 600, distance_limit = -1,
  columns = c("longitude", "latitude", "time", "distance"),
  reverse_flow = FALSE, profile = "car")
}
\arguments{
\item{start_point}{The start point as (lon, lat) pair.}

\item{time_limit}{The travel time limit in seconds.
Ignored if \code{distance_limit > 0}.}

\item{distance_limit}{The distance limit in meters.}

\item{columns}{The columns that should be returned.}

\item{reverse_flow}{Use \code{reverse_flow = TRUE} to change the flow direction.}

\item{profile}{The profile for which the spt should be calculated.}

\item{See}{\link{gh_available_spt_columns} for available columns.}
}
\description{
Get the shortest path tree for a given start point
}
