% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-route.R
\name{gh_instructions}
\alias{gh_instructions}
\title{Extract the instructions from a gh route object}
\usage{
gh_instructions(data)
}
\arguments{
\item{data}{A \code{gh_route} object.}
}
\description{
Extract the instructions from a gh route object
}
\seealso{
\link{gh_get_route}
}
