% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-route.R
\name{gh_points}
\alias{gh_points}
\title{Extract the points from a gh route object}
\usage{
gh_points(data)
}
\arguments{
\item{data}{A \code{gh_route} object.}
}
\description{
Extract the points from a gh route object
}
