% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{gh_route_coordinates}
\alias{gh_route_coordinates}
\title{Get coordinates from route.}
\usage{
gh_route_coordinates(route)
}
\arguments{
\item{route}{route (raw) route object, see \code{\link{gh_get_route}}}
}
\description{
Get coordinates from route.
}
