% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{gh_route_linestring}
\alias{gh_route_linestring}
\title{Parse route to linestring.}
\usage{
gh_route_linestring(route)
}
\arguments{
\item{route}{(raw) route object, see \code{\link{gh_get_route}}}
}
\description{
Parse route to linestring.
}
