% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{gh_set_api_url}
\alias{gh_set_api_url}
\title{Set GH-API base url.}
\usage{
gh_set_api_url(api_url)
}
\arguments{
\item{api_url}{API base url}
}
\description{
Set GH-API base url.
}
\note{
Internally it calls \code{Sys.setenv} to store the API url
in an environment variable called \code{GH_API_URL}.
}
