% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt-endpoint.R
\name{gh_spt_as_linestrings_sf}
\alias{gh_spt_as_linestrings_sf}
\title{Build lines from gh spt opbject}
\usage{
gh_spt_as_linestrings_sf(data)
}
\arguments{
\item{data}{A \code{gh_spt} object.}
}
\description{
Build lines from gh spt opbject
}
