% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt-endpoint.R
\name{gh_spt_columns}
\alias{gh_spt_columns}
\title{Select the columns to be returned by a spt request}
\usage{
gh_spt_columns(longitude = TRUE, latitude = TRUE, time = TRUE,
  distance = TRUE, prev_longitude = FALSE, prev_latitude = FALSE,
  prev_time = FALSE, prev_distance = FALSE, node_id = FALSE,
  prev_node_id = FALSE, edge_id = FALSE, prev_edge_id = FALSE)
}
\arguments{
\item{longitude, latitude}{The longitude, latitude of the node.}

\item{time, distance}{The travel time, distance to the node.}

\item{prev_longitude, prev_latitude}{The longitude, latitude of the previous node.}

\item{prev_time, prev_distance}{The travel time, distance to the previous node.}

\item{node_id, prev_node_id}{The ID of the node, previous node.}

\item{edge_id, prev_edge_id}{The ID of the edge, previous edge.}
}
\description{
Times are returned in milliseconds and distances in meters.
}
