% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-route.R
\name{gh_time_distance}
\alias{gh_time_distance}
\title{Extract time and distance from a gh route object}
\usage{
gh_time_distance(data)
}
\arguments{
\item{data}{A \code{gh_route} object.}
}
\description{
Extract time and distance from a gh route object
}
