% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone-api.R
\name{isochrone_get}
\alias{isochrone_get}
\title{Get the isochrones for a given point.}
\usage{
isochrone_get(api_client, point, ...)
}
\arguments{
\item{api_client}{api client, see \code{\link{api_client}}}

\item{point}{numeric vector; [lat, lng] pair}

\item{...}{additional query parameters, see \url{https://docs.graphhopper.com}}
}
\description{
Get the isochrones for a given point.
}
