% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route-api.R
\name{route_get}
\alias{route_get}
\title{Get the route for a given set of points.}
\usage{
route_get(api_client, points, ..., sf = TRUE)
}
\arguments{
\item{api_client}{api client, see \code{\link{api_client}}}

\item{points}{list of points as [lat, lng] pairs}

\item{...}{additional query parameters, see \url{https://docs.graphhopper.com}}

\item{sf}{logical scalar; whether to return route also as \code{sf} object}
}
\description{
Get the route for a given set of points.
}
\examples{
\dontrun{
   start_point <- c(52.592204, 13.414307)
   end_point <- c(52.539614, 13.364868)

   route_get(list(start_point, end_point), locale = "de")
}
}
