% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-endpoint.R
\name{gh_get_info}
\alias{gh_get_info}
\title{Get information about the GraphHopper instance}
\usage{
gh_get_info()
}
\description{
Get information about the GraphHopper instance
}
\examples{
\dontrun{
info <- gh_get_info()

message(info$version)
message(info$data_date)
print(gh_bbox(info))
}
}
