% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route-endpoint.R
\name{gh_get_route}
\alias{gh_get_route}
\title{Get a route for a given set of points}
\usage{
gh_get_route(points, ..., response_only = FALSE)
}
\arguments{
\item{points}{A list of 2 or more points as (lat, lon) pairs.}

\item{...}{Optional parameters that are passed to the query.}

\item{response_only}{Whether to return the raw response object instead of
just its content.}
}
\description{
Get a route for a given set of points
}
\seealso{
\url{https://docs.graphhopper.com/#tag/Routing-API} for optional parameters.
}
