% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{gh_get_route}
\alias{gh_get_route}
\title{Get the route for a given set of points.}
\usage{
gh_get_route(points, ...)
}
\arguments{
\item{points}{list of points as [lat, lng] pairs}

\item{...}{additional query parameters, see \url{https://docs.graphhopper.com}}
}
\description{
Get the route for a given set of points.
}
