% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt-endpoint.R
\name{gh_get_spt}
\alias{gh_get_spt}
\title{Get the shortest path tree for a given start point}
\usage{
gh_get_spt(start_point, time_limit = 600, columns = c("longitude",
  "latitude", "time", "distance", "prev_longitude", "prev_latitude"))
}
\arguments{
\item{start_point}{The start point as (lon, lat) pair.}

\item{time_limit}{The travel time limit in seconds.}

\item{columns}{The columns that should be returned.
See \link{gh_available_spt_columns} for available columns.}
}
\description{
Get the shortest path tree for a given start point
}
